+++ canopen_analyzer/dox/documentation/html/classcanopen__sniffer_1_1canopen__sniffer.js
@@ -1,16 +1,18 @@
 var classcanopen__sniffer_1_1canopen__sniffer =
 [
-    [ "__init__", "classcanopen__sniffer_1_1canopen__sniffer.html#a8bbb50839e753a6eb1acae46c7892335", null ],
+    [ "__init__", "classcanopen__sniffer_1_1canopen__sniffer.html#a0dedd9047111690fed6d3ffc28db4da6", null ],
     [ "_dispatch_request", "classcanopen__sniffer_1_1canopen__sniffer.html#a289e8d72a57bec3cc07a63a89991bd25", null ],
     [ "_ensure_bus", "classcanopen__sniffer_1_1canopen__sniffer.html#a6a3ccdcf3682f03c4ed720908318733e", null ],
     [ "_handle_requested_frame", "classcanopen__sniffer_1_1canopen__sniffer.html#a56691f45b9cd4ae8a01630a8cf372c01", null ],
+    [ "_json_safe_raw_frame", "classcanopen__sniffer_1_1canopen__sniffer.html#a1554dc53d58b5b1b482367d3ca6b87e5", null ],
+    [ "export_raw_frame", "classcanopen__sniffer_1_1canopen__sniffer.html#a7bee1e0390714d04ec48916af56fee49", null ],
     [ "handle_received_message", "classcanopen__sniffer_1_1canopen__sniffer.html#a71281faea956cdc27e9f39548355a5a9", null ],
     [ "run", "classcanopen__sniffer_1_1canopen__sniffer.html#ad22709b2e67308af35f55680d5a026e0", null ],
-    [ "save_frame_to_csv", "classcanopen__sniffer_1_1canopen__sniffer.html#af7adf7c35981de206cae05750cfdeb0f", null ],
     [ "send_raw_pdo", "classcanopen__sniffer_1_1canopen__sniffer.html#af5700ac79e73b5f715f582e443ccc969", null ],
     [ "send_sdo_download", "classcanopen__sniffer_1_1canopen__sniffer.html#aa8d11ec8573e1db81498322bba468a2d", null ],
     [ "send_sdo_upload_request", "classcanopen__sniffer_1_1canopen__sniffer.html#a097369338fe5940560a0d109f2416bc5", null ],
     [ "stop", "classcanopen__sniffer_1_1canopen__sniffer.html#a2342b933b01f118be030e93341908480", null ],
+    [ "_json_first", "classcanopen__sniffer_1_1canopen__sniffer.html#af086836ec2a067957f99fe7b6edf8fc5", null ],
     [ "_stop_event", "classcanopen__sniffer_1_1canopen__sniffer.html#a2c9373a2aa1f4cdf3e6c936e81b17267", null ],
     [ "bus", "classcanopen__sniffer_1_1canopen__sniffer.html#a411bd59fecc82fdfda0e5c9d7da7447f", null ],
     [ "export", "classcanopen__sniffer_1_1canopen__sniffer.html#af44e2d3afd4e0e4d03f637dbb2bce3cd", null ],
@@ -21,6 +23,7 @@ var classcanopen__sniffer_1_1canopen__sniffer =
     [ "interface", "classcanopen__sniffer_1_1canopen__sniffer.html#ae6deea6e7d03b5d4db98ea209159e57f", null ],
     [ "log", "classcanopen__sniffer_1_1canopen__sniffer.html#ab09a63eb35b270b5cdbead1983ebdccb", null ],
     [ "network", "classcanopen__sniffer_1_1canopen__sniffer.html#a6b288a470bcb2fd5f321915ef4045b8b", null ],
+    [ "pcap_writer", "classcanopen__sniffer_1_1canopen__sniffer.html#a6ffee8a5264219b0fb83d1d87b6bdbd6", null ],
     [ "raw_frame", "classcanopen__sniffer_1_1canopen__sniffer.html#a796de4da9ed5fad20bff2ba1ea3b7051", null ],
     [ "requested_frame", "classcanopen__sniffer_1_1canopen__sniffer.html#ab4a622f1e1275087c3558661cb25300a", null ]
 ];