+++ canopen_analyzer/dox/documentation/html/classprocess__frames_1_1process__frames.js
@@ -1,12 +1,13 @@
 var classprocess__frames_1_1process__frames =
 [
-    [ "__init__", "classprocess__frames_1_1process__frames.html#ae9017521a9672251ad6a11dcd1d6103d", null ],
+    [ "__init__", "classprocess__frames_1_1process__frames.html#a5bc12514c98edca8798862652dd6bd4d", null ],
+    [ "_json_safe_processed_frame", "classprocess__frames_1_1process__frames.html#a4c7031035c06abd316cf1ae48aee76a7", null ],
     [ "_sdo_has_index", "classprocess__frames_1_1process__frames.html#a2c327b400fe52a2eebc76e7c41123b2a", null ],
+    [ "export_processed_frame", "classprocess__frames_1_1process__frames.html#aa8b1ed89c91c825feda35358296d65f4", null ],
     [ "run", "classprocess__frames_1_1process__frames.html#ad22709b2e67308af35f55680d5a026e0", null ],
-    [ "save_frame", "classprocess__frames_1_1process__frames.html#a978c143f89b0822beb716185289b7bcf", null ],
-    [ "save_frame_to_csv", "classprocess__frames_1_1process__frames.html#aa9acd531bcaa66901bfc31865c7de1d3", null ],
-    [ "save_processed_frame", "classprocess__frames_1_1process__frames.html#af72c85667f791c7f82aa7b89ac4acf11", null ],
+    [ "save_processed_frame", "classprocess__frames_1_1process__frames.html#abab717783b0e365598b1f9c9205577d2", null ],
     [ "stop", "classprocess__frames_1_1process__frames.html#a26ca7c1c7fcdd35378e7be97727047a6", null ],
+    [ "_json_first", "classprocess__frames_1_1process__frames.html#af086836ec2a067957f99fe7b6edf8fc5", null ],
     [ "_stop_event", "classprocess__frames_1_1process__frames.html#a2c9373a2aa1f4cdf3e6c936e81b17267", null ],
     [ "eds_map", "classprocess__frames_1_1process__frames.html#a087b69c6d3cd89b6872e63605be5e05c", null ],
     [ "export", "classprocess__frames_1_1process__frames.html#af44e2d3afd4e0e4d03f637dbb2bce3cd", null ],